Create Project: Kronos To Do Tracking
This project was created as a way for HR to efficiently track outstanding checklist items for employees since the built-in report in Kronos doesn't allow for filtering on any aspect specific to the to do (checklist name, item name, etc.). It is intended to be saved as a macro-enabled workbook (.xltm) so that each time it is opened a blank copy of itself is created. It will create a menu item on the Add-ins ribbon under called "Track Kronos To Dos" under the item "Custom Workbook Functions." The idea is to copy/paste the entire page directly form the Kronos website (CTRL + A; CTRL + C; CTRL + V). Each entire page of the "To Do Tracking" saved view of the "All To Do Items" report in Kronos should be copy/pasted into its own sheet. The only sheets that should exist in the workbook should be those containing the copy/pasted full page from the Kronos website. After all data is copy/pasted, the "Track Kronos To Dos" menu item should be clicked. This will combine all of the data from each of the copied pages into its own "Data" worksheet at the beginning of all the worksheets. Then two worksheets will be created after this with pivot tables on them to analyze the data by checklist or approver.